feat: DB function for storing new known sys errors
To avoid sending repeated error emails to the same teachers about the
same issues, when sending an email we'll store the details of the issue
that we're notifying the teacher about in the knownsyserrors table.

This commit adds a helper function for storing these.

Contributes to: #1
